[refactor] 구조적으로 변경
